userapi/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-CORS
Flask-Bcrypt
python-dotenv
pytest
gunicorn
flasgger